Restructures the html.
I've moved some of the sentance areound. Updated the styling
to be more consistent with the rest of the page.

As part of it I thought it would make sense to move FAQ to
mentorship as its own section and remove it from the main page.

Fixes linting issues with FAQ.js